Allow configuring the interface for the individual services
The interfaces for the services (http, rpc and serf) in nomad can be
configured individually so that they can appear on different networks.

For instance, one might want to generally use the localnet of the cloud
project for communication between nomad nodes, but one would like the http
service availabe over tailscale.

This change allows for this configuration flexibility to be exposed in the
nomad role.